requirements.txt
bitstring>=3.1.5
cinemagoer>=2022.12.27
colorama>=0.3.9
libprick>=1.1.0
mutagen>=1.40.0
numpy==1.19.0
opencv-python-headless>=3.4.0.12
pymediainfo>=2.2.1
PyInstaller>=3.4
pypiwin32>=223 ; sys_platform == 'win32'
PySimpleGUI==4.32.1
python-magic>=0.4.15
python-magic-bin>=0.4.14 ; sys_platform == 'win32' or sys_platform == 'darwin'
requests>=2.18.4
release-dir-scanner>=1.0.1
termcolor>=1.1.0